Goalies behave intelligently
Goalies move towards the ball if is near their goal
Goalies move in between the ball and their goal if the ball is in their half of the pitch.
Goalies move to a point in front of their goal if the ball is in the opponents half of the pitch.